Configures cd command to only run when batchMode is false
It doesn’t make sense for cd command to run from a batch file because
the batch file may reference a directory that does not exist on the
computer that it is being run on.

At least it seems that way to me. If someone thinks otherwise we can
discuss it and change it if needed.